added welcome page and update on delete function

+++ project/customer_delete.php
@@ -1,11 +1,27 @@
 <?php
-    // include database connection
-    include 'config/database.php';
-    try {
-        // get record ID
-        // isset() is a PHP function used to verify if a value is there or not
-        $user_id = isset($_GET['user_id']) ? $_GET['user_id'] :  die('ERROR: Record ID not found.');
+// include database connection
+include 'config/database.php';
+try {
+    // get record ID
+    // isset() is a PHP function used to verify if a value is there or not
+    $user_id = isset($_GET['user_id']) ? $_GET['user_id'] :  die('ERROR: Record ID not found.');
 
+    $select = "SELECT username AS check_user FROM customers WHERE user_id=:user_id";
+    $stmt = $con->prepare($select);
+    $stmt->bindParam(":user_id", $user_id);
+    $stmt->execute();
+    $row = $stmt->fetch(PDO::FETCH_ASSOC);
+    extract($row);
+
+    $check = "SELECT username FROM order_summary WHERE username=:username";
+    $stmt = $con->prepare($check);
+    $stmt->bindParam(":username", $check_user);
+    $stmt->execute();
+    $count = $stmt->rowCount();
+
+    if ($count > 0) {
+        echo "This customer had order placed so cannot be delete.";
+    } else {
         // delete query
         $query = "DELETE FROM customers WHERE user_id = ?";
         $stmt = $con->prepare($query);
@@ -19,7 +35,8 @@
             die('Unable to delete record.');
         }
     }
-    // show error
-    catch (PDOException $exception) {
-        die('ERROR: ' . $exception->getMessage());
-    }
+}
+// show error
+catch (PDOException $exception) {
+    die('ERROR: ' . $exception->getMessage());
+}
