+++ project/product_delete.php
@@ -1,11 +1,27 @@
 <?php
-    // include database connection
-    include 'config/database.php';
-    try {
-        // get record ID
-        // isset() is a PHP function used to verify if a value is there or not
-        $id = isset($_GET['id']) ? $_GET['id'] :  die('ERROR: Record ID not found.');
+// include database connection
+include 'config/database.php';
+try {
+    // get record ID
+    // isset() is a PHP function used to verify if a value is there or not
+    $id = isset($_GET['id']) ? $_GET['id'] :  die('ERROR: Record ID not found.');
 
+    $select = "SELECT id AS check_product FROM products WHERE id=:id";
+    $stmt = $con->prepare($select);
+    $stmt->bindParam(":id", $id);
+    $stmt->execute();
+    $row = $stmt->fetch(PDO::FETCH_ASSOC);
+    extract($row);
+
+    $check = "SELECT product_id FROM order_details WHERE product_id=:product_id";
+    $stmt = $con->prepare($check);
+    $stmt->bindParam(":product_id", $check_product);
+    $stmt->execute();
+    $count = $stmt->rowCount();
+
+    if ($count > 0) {
+        echo "This product has been ordered so cannot be delete.";
+    } else {
         // delete query
         $query = "DELETE FROM products WHERE id = ?";
         $stmt = $con->prepare($query);
@@ -19,7 +35,8 @@
             die('Unable to delete record.');
         }
     }
-    // show error
-    catch (PDOException $exception) {
-        die('ERROR: ' . $exception->getMessage());
-    }
+}
+// show error
+catch (PDOException $exception) {
+    die('ERROR: ' . $exception->getMessage());
+}
